@@ -35,36 +35,39 @@
     "gulp": "gulp"
   },
   "dependencies": {
+    "@angular/cli": "^10.0.5",
+    "@angular/compiler-cli": "^10.0.8",
     "@types/adal": "1.0.29",
-    "adal-angular": "1.0.15"
+    "adal-angular": "1.0.17",
+    "tslib": "2.0.0"
   },
-  "peerDependencies": {
-    "@angular/core": ">= 5.0.0"
-  },
+  "peerDependencies": {},
   "devDependencies": {
-    "@angular/common": "5.1.0",
-    "@angular/compiler": "5.1.0",
-    "@angular/core": "5.1.0",
-    "@angular/platform-browser": "5.1.0",
-    "del": "^2.2.0",
+    "@angular-devkit/build-angular": "~0.1000.5",
+    "@angular/compiler": "10.0.8",
+    "@angular/common": "10.0.8",
+    "@angular/core": "10.0.8",
+    "@angular/platform-browser": "10.0.8",
+    "codelyzer": "^6.0.0",
+    "del": "^5.1.0",
     "gulp": "^3.9.1",
-    "gulp-git": "^1.8.0",
-    "gulp-header": "^1.8.2",
-    "gulp-load-plugins": "^1.2.4",
-    "gulp-rename": "^1.2.2",
-    "gulp-resolve-dependencies": "^2.2.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-template": "^4.0.0",
-    "gulp-typescript": "^3.1.3",
+    "gulp-git": "^2.10.1",
+    "gulp-header": "^2.0.9",
+    "gulp-load-plugins": "^2.0.3",
+    "gulp-rename": "^2.0.0",
+    "gulp-resolve-dependencies": "^3.0.1",
+    "gulp-sourcemaps": "^2.6.5",
+    "gulp-template": "^5.0.0",
+    "gulp-typescript": "^5.0.1",
     "gulp-util": "^3.0.8",
-    "merge2": "^1.0.2",
-    "reflect-metadata": "0.1.10",
-    "require-dir": "^0.3.0",
-    "run-sequence": "^1.2.1",
-    "rxjs": "5.5.2",
-    "systemjs-builder": "^0.16.3",
-    "typescript": "2.4.2",
-    "typings": "^1.5.0",
-    "zone.js": "0.8.18"
+    "merge2": "^1.4.1",
+    "reflect-metadata": "0.1.13",
+    "require-dir": "^1.2.0",
+    "run-sequence": "^2.2.1",
+    "rxjs": "6.6.2",
+    "systemjs-builder": "^0.16.15",
+    "typescript": "3.9.7",
+    "typings": "^2.1.1",
+    "zone.js": "~0.10.3"
   }
 }
